reverted back to 16 pixels per bank and passed bank to SpecBankCal

+++ lmtslr/viewer/spec_viewer.py
@@ -605,7 +605,7 @@ class SpecCalViewer(SpecViewer):
             plot_scale = plot_scale / nscale
         plot_order = [1,5,9,13,2,6,10,14,3,7,11,15,4,8,12,16];
         print(S.roach_pixel_ids)
-        nrows = S.npix/4
+        nrows = int(S.npix/4)
         for ipix in range(S.npix):
             pixel_id = S.roach_pixel_ids[ipix]
             if nrows == 1:
